+++ plugins/hostinger/vue-frontend/src/App.vue
@@ -3,14 +3,24 @@ import { RouterView } from "vue-router";
 import Modals from "@/components/Modals/Base/Modals.vue";
 import Wrapper from "@/layouts/Wrapper.vue";
 import { useRoute } from "vue-router";
+import { computed } from "vue";
+import { HeaderButton } from "@/types";
 
 const route = useRoute();
+
+const headerTitle = computed(() => {
+  return route.meta.title as string;
+});
+
+const headerButton = computed(() => {
+  return route.meta.headerButton as HeaderButton;
+});
 </script>
 
 <template>
   <div>
     <div id="overhead-button" />
-    <Wrapper :title="route.meta.title">
+    <Wrapper :title="headerTitle" :header-button="headerButton">
       <RouterView v-slot="{ Component }">
         <Component :is="Component" />
       </RouterView>