+++ plugins/hostinger/vue-frontend/src/types/models/generalDataModels.ts
@@ -17,6 +17,7 @@ export type HostingerToolsData = {
   homeUrl: string;
   siteUrl: string;
   pluginUrl: string;
+  assetUrl: string;
   translations: { [key: string]: string };
   restBaseUrl: string;
   nonce: string;